refactor: 기능 단위로 테스트 파일 분리
